[TD-597] 사장님 계정 회원탈퇴 API 추가
