+++ threedollar-api-boss/src/main/kotlin/com/depromeet/threedollar/boss/api/controller/auth/AuthController.kt
@@ -2,13 +2,15 @@ package com.depromeet.threedollar.boss.api.controller.auth
 
 import com.depromeet.threedollar.application.common.dto.ApiResponse
 import com.depromeet.threedollar.boss.api.config.interceptor.Auth
+import com.depromeet.threedollar.boss.api.config.resolver.BossId
 import com.depromeet.threedollar.boss.api.config.session.SessionConstants.BOSS_ACCOUNT_ID
+import com.depromeet.threedollar.boss.api.service.auth.AuthCommonService
 import com.depromeet.threedollar.boss.api.service.auth.AuthServiceProvider
 import com.depromeet.threedollar.boss.api.service.auth.dto.request.LoginRequest
-import com.depromeet.threedollar.boss.api.service.auth.dto.response.LoginResponse
-import com.depromeet.threedollar.boss.api.service.auth.SignupService
 import com.depromeet.threedollar.boss.api.service.auth.dto.request.SignupRequest
+import com.depromeet.threedollar.boss.api.service.auth.dto.response.LoginResponse
 import io.swagger.annotations.ApiOperation
+import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RestController
@@ -19,17 +21,17 @@ import javax.validation.Valid
 class AuthController(
     private val httpSession: HttpSession,
     private val authServiceProvider: AuthServiceProvider,
-    private val signupService: SignupService
+    private val authCommonService: AuthCommonService
 ) {
 
     @ApiOperation("사장님 계정의 회원가입을 요청합니다. (차후 승인이 필요합니다)")
     @PostMapping("/v1/auth/signup")
-    fun signup(
+    fun signUp(
         @Valid @RequestBody request: SignupRequest
     ): ApiResponse<String> {
         val authService = authServiceProvider.getAuthService(request.socialType)
         val socialId = authService.getSocialId(LoginRequest(request.token, request.socialType))
-        signupService.signup(request, socialId)
+        authCommonService.signUp(request, socialId)
         return ApiResponse.SUCCESS
     }
 
@@ -49,6 +51,17 @@ class AuthController(
     @PostMapping("/v1/auth/logout")
     fun logout(): ApiResponse<String> {
         httpSession.removeAttribute(BOSS_ACCOUNT_ID)
+        return ApiResponse.SUCCESS
+    }
+
+    @ApiOperation("[인증] 사장님 계정을 회원탈퇴 합니다")
+    @Auth
+    @DeleteMapping("/v1/auth/sign-out")
+    fun signOut(
+        @BossId bossId: String
+    ): ApiResponse<String> {
+        authCommonService.signOut(bossId)
+        httpSession.invalidate()
         return ApiResponse.SUCCESS
     }
