+++ threedollar-api-boss/src/main/kotlin/com/depromeet/threedollar/boss/api/service/auth/AuthCommonService.kt
@@ -9,20 +9,23 @@ import com.depromeet.threedollar.document.boss.document.account.BossAccountRepos
 import com.depromeet.threedollar.document.boss.document.account.BossAccountSocialType
 import com.depromeet.threedollar.document.boss.document.category.BossStoreCategoryRepository
 import com.depromeet.threedollar.document.boss.document.registration.RegistrationRepository
+import com.depromeet.threedollar.document.boss.document.withdrawal.BossWithdrawalAccount
+import com.depromeet.threedollar.document.boss.document.withdrawal.BossWithdrawalAccountRepository
 import org.springframework.context.ApplicationEventPublisher
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
 @Service
-class SignupService(
+class AuthCommonService(
     private val bossAccountRepository: BossAccountRepository,
+    private val bossWithdrawalAccountRepository: BossWithdrawalAccountRepository,
     private val registrationRepository: RegistrationRepository,
     private val bossStoreCategoryRepository: BossStoreCategoryRepository,
     private val eventPublisher: ApplicationEventPublisher
 ) {
 
     @Transactional
-    fun signup(request: SignupRequest, socialId: String) {
+    fun signUp(request: SignupRequest, socialId: String) {
         validateDuplicateRegistration(socialId, request.socialType)
         BossStoreCategoryServiceUtils.validateExistsCategories(bossStoreCategoryRepository, request.storeCategoriesIds)
         val registration = request.toEntity(socialId)
@@ -36,6 +39,13 @@ class SignupService(
         if (registrationRepository.existsRegistrationBySocialIdAndSocialType(socialId, socialType)) {
             throw ForbiddenException("가입 신청 후 대기중인 사장님(${socialId} - (${socialType}) 입니다.", ErrorCode.FORBIDDEN_WAITING_APPROVE_BOSS_ACCOUNT)
         }
+    }
+
+    @Transactional
+    fun signOut(bossId: String) {
+        val bossAccount = BossAccountServiceUtils.findBossAccountById(bossAccountRepository, bossId)
+        bossWithdrawalAccountRepository.save(BossWithdrawalAccount.newInstance(bossAccount))
+        bossAccountRepository.delete(bossAccount)
     }
 
 }
