+++ threedollar-document/src/main/kotlin/com/depromeet/threedollar/document/boss/document/account/BossAccount.kt
@@ -27,7 +27,7 @@ class BossAccount(
         ): BossAccount {
             return BossAccount(
                 name = name,
-                socialInfo = BossAccountSocialInfo(socialId, socialType),
+                socialInfo = BossAccountSocialInfo.of(socialId, socialType),
                 businessNumber = businessNumber,
                 pushSettingsStatus = pushSettingsStatus
             )
