+++ threedollar-document/src/main/kotlin/com/depromeet/threedollar/document/boss/document/account/BossAccountSocialInfo.kt
@@ -3,4 +3,15 @@ package com.depromeet.threedollar.document.boss.document.account
 data class BossAccountSocialInfo(
     val socialId: String,
     val socialType: BossAccountSocialType
-)
+) {
+
+    companion object {
+        fun of(socialId: String, socialType: BossAccountSocialType): BossAccountSocialInfo {
+            return BossAccountSocialInfo(
+                socialId = socialId,
+                socialType = socialType
+            )
+        }
+    }
+
+}
