+++ threedollar-document/src/main/kotlin/com/depromeet/threedollar/document/boss/document/withdrawal/BossWithdrawalAccount.kt
@@ -1,5 +1,6 @@
 package com.depromeet.threedollar.document.boss.document.withdrawal
 
+import com.depromeet.threedollar.document.boss.document.account.BossAccount
 import com.depromeet.threedollar.document.boss.document.account.BossAccountSocialInfo
 import com.depromeet.threedollar.document.boss.document.account.PushSettingsStatus
 import com.depromeet.threedollar.document.common.document.BaseDocument
@@ -14,10 +15,35 @@ class BossWithdrawalAccount(
     val socialInfo: BossAccountSocialInfo,
     val businessNumber: BusinessNumber,
     val pushSettingsStatus: PushSettingsStatus,
-) : BaseDocument()
+) : BaseDocument() {
+
+    companion object {
+        fun newInstance(bossAccount: BossAccount): BossWithdrawalAccount {
+            return BossWithdrawalAccount(
+                backupInfo = BackupBossAccountInfo.of(bossAccount),
+                name = bossAccount.name,
+                socialInfo = bossAccount.socialInfo.copy(),
+                businessNumber = bossAccount.businessNumber.copy(),
+                pushSettingsStatus = bossAccount.pushSettingsStatus
+            )
+        }
+    }
+
+}
 
 
 data class BackupBossAccountInfo(
     val bossId: String,
     val bossCreatedDateTime: LocalDateTime
-)
+) {
+
+    companion object {
+        fun of(bossAccount: BossAccount): BackupBossAccountInfo {
+            return BackupBossAccountInfo(
+                bossId = bossAccount.id,
+                bossCreatedDateTime = bossAccount.createdDateTime
+            )
+        }
+    }
+
+}
