+++ threedollar-document/src/testFixtures/kotlin/com/depromeet/threedollar/document/boss/document/account/BossAccountCreator.kt
@@ -15,7 +15,7 @@ object BossAccountCreator {
     ): BossAccount {
         return BossAccount(
             name = name,
-            socialInfo = BossAccountSocialInfo(socialId, socialType),
+            socialInfo = BossAccountSocialInfo.of(socialId, socialType),
             businessNumber = businessNumber,
             pushSettingsStatus = pushSettingsStatus
         )
